requirements.txt
# Telegram Bot
aiogram==3.15.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# AI APIs
anthropic==0.42.0
openai==1.58.1

# Scheduling
apscheduler==4.0.0a5

# Configuration
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Visualization
matplotlib==3.10.0

# Presentation generation
python-pptx==1.0.2

# Logging
loguru==0.7.3

# HTTP (for Whisper file downloads)
aiohttp==3.11.11
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
testcontainers[postgres]==4.9.0
